+++ README.md
@@ -69,7 +69,7 @@ Prompts are input messages or queries you send to a Large Language model. You ca
 Python
 
 ```python
-from aiconfig.AIConfigSettings import ModelMetadata, Prompt
+from aiconfig.schema import ModelMetadata, Prompt
 
 prompt = Prompt(
     name="prompt1",
