+++ python/src/aiconfig/schema.py
