+++ python/tests/test_dataclass.py
@@ -1,5 +1,5 @@
 
-from aiconfig.AIConfigSettings import ModelMetadata, Prompt, PromptInput, PromptMetadata
+from aiconfig.schema import ModelMetadata, Prompt, PromptInput, PromptMetadata
 
 
 def test_get_raw_prompt_from_prompt_object_string_input():
