+++ python/tests/test_util/test_params.py
@@ -1,4 +1,4 @@
-from aiconfig.AIConfigSettings import Prompt, PromptMetadata
+from aiconfig.schema import Prompt, PromptMetadata
 from aiconfig.util.params import find_dependencies_in_prompt, get_dependency_graph, get_parameters_in_template
 import pytest
